% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extend_oag}
\alias{extend_oag}
\title{Extrapolating Open Age group}
\usage{
extend_oag(pop, country, year = 2023, oag_new = 100, n = 1)
}
\arguments{
\item{pop}{Data table of population counts with columns 
\dQuote{age}, \dQuote{popM} and \dQuote{popF}. Ages can be either 
single year or five year age groups.}

\item{country}{Name of the country to use for the population standard.}

\item{year}{Which year should be used for extracting the standard population.}

\item{oag_new}{Integer defining the new OAG. It must be available in WPP population datasets,
i.e. it must be smaller equal 100. If \code{pop} is defined for five year age groups,
\code{oag_new} must be a multiple of 5.}

\item{n}{Size of age groups. It can be either 1 (default) or 5. It must correspond 
to the age grouping in the \code{pop} dataset.}
}
\value{
Data table with the same columns as in the \code{pop} dataset, 
    where the original last age group is dissaggregated up to the age 
    given by \code{oag_new}.
}
\description{
Extrapolate the open age group (OAG) into multiple ages using a standard population.
}
\examples{
# extract population of Germany in 2023 
pop_full <- get_wpp_pop("Germany", year = 2023)

# collapse the OAG into 85
pop_oag85 <- reduce_oag(pop_full, 85)

# extrapolate to 95
pop_oag95 <- extend_oag(pop_oag85, "Germany", year = 2023, oag_new = 95)

# plot results
plot(pop_oag85[, age], pop_oag85[, popM], type = "l", xlim = c(50, 100),
    ylim = range(pop_oag85[, popM], pop_full[, popM]),
    main = "Male Population of Germany in 2023", col = "blue",
    xlab = "age", ylab = "Population (in thousands)")
lines(pop_full[, age], pop_full[, popM], col = "black")
lines(pop_oag95[, age], pop_oag95[, popM], col = "red")
legend("bottomleft", legend = c("full", "OAG = 85", "OAG = 95"),
    bty = "n", lty = 1, col = c("black", "blue", "red"))

# check the sums for pop of single year ages
all.equal(pop_full[, sum(popF + popM)], pop_oag85[, sum(popF + popM)])
all.equal(pop_full[, sum(popF + popM)], pop_oag95[, sum(popF + popM)])

# 5-year age groups
pop5_full <- get_wpp_pop("Germany", year = 2023, n = 5)
pop5_oag80 <- reduce_oag(pop5_full, 80)
pop5_oag100 <- extend_oag(pop5_oag80, "Germany", year = 2023, oag_new = 100, n = 5)
# check the sums
all.equal(pop5_full[, sum(popF + popM)], pop5_oag80[, sum(popF + popM)])
all.equal(pop5_full[, sum(popF + popM)], pop5_oag100[, sum(popF + popM)])

}
