% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_targeted_measure}
\alias{get_targeted_measure}
\title{Compute Targeted Measure}
\usage{
get_targeted_measure(
  target,
  dt,
  start_year = NULL,
  end_year = NULL,
  measure = "tfr"
)
}
\arguments{
\item{target}{Targeted value at \code{end_year}.}

\item{dt}{Data table containing values to be modified by the interpolated values. It is expected to 
have a column "year" and another column containing the actual measure. Its name must correspond
to the \code{measure} argument, by default "tfr".}

\item{start_year}{Year to start the interpolation.}

\item{end_year}{Year at which the target value is to be reached.}

\item{measure}{Name of the column to be interpolated.}
}
\description{
Helper function to obtain targeted measure, interpolated between start and end year.
}
\examples{
# Change the TFR of Niger to reach 4.0 at 2100, starting from 2024
my_tfr <- get_wpp_tfr("Niger")
my_tfr_mod <- get_targeted_measure(4, my_tfr, start_year = 2024)

# plot the difference
plot(my_tfr$year, my_tfr$tfr, type = "l", xlab = "", ylab = "TFR", main = "Niger")
lines(my_tfr_mod$year, my_tfr_mod$tfr, col = "red")

}
