% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpp.R
\name{get_wpp}
\alias{get_wpp}
\title{Getter function for the WPP datasets.}
\usage{
get_wpp(dataset_name, package = "wpp2022")
}
\description{
Since some of the datasets are big, this function caches already used datasets.
}
\keyword{internal}
