% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpp.R
\name{get_wpp_countries}
\alias{get_wpp_countries}
\title{WPP countries}
\usage{
get_wpp_countries(include_aggregates = FALSE)
}
\arguments{
\item{include_aggregates}{Logical. If \code{TRUE} the set of countries includes
aggregated regions, such as continents or other aggregations of countries.
If \code{FALSE} (default) only countries are returned.}
}
\value{
Character vector of names of countries and optionally aggregations.
}
\description{
Returns a vector of country names included in the WPP data
}
\details{
The function uses the \code{\link[wpp2022]{UNlocations}} dataset of the
     \pkg{wpp2022} package. By default it selects locations where \code{location_type} is 4.
     If \code{include_aggregates} is \code{TRUE} it returns all locations from this dataset.
}
\examples{
# Countries that start with "E"
Ecountries <- grep("^E", get_wpp_countries(), value = TRUE)
print(Ecountries)
# Europe should not be there
"Europe" \%in\% Ecountries # should be FALSE

# include aggregations
Ecountries2 <- grep("^E", get_wpp_countries(include_aggregates = TRUE),
    value = TRUE)
print(Ecountries2)
"Europe" \%in\% Ecountries2 # should be TRUE as Europe is included

}
