% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpp.R
\name{get_wpp_e0}
\alias{get_wpp_e0}
\title{Country-specific WPP life expectancy at birth}
\usage{
get_wpp_e0(country)
}
\arguments{
\item{country}{Name of country.}
}
\value{
\code{data.table} object with columns \code{year}, \code{e0M} and \code{e0F}.
}
\description{
The function returns the life expectancy at birth for male and female over time for a specific country
      from the UN's World Population Prospects.
}
\details{
The data is extracted by combining datasets \code{\link[wpp2022]{e01dt}} and
     \code{\link[wpp2022]{e0proj1dt}}.
}
\examples{
# Compare female e0 of Japan and Norway
e0_norway <- get_wpp_e0("Norway")
e0_japan <- get_wpp_e0("Japan")
plot(e0_japan[, year], e0_japan[, e0F], type = "l", ylim = c(50, 100),
    main = "Female e0 of Japan and Norway", xlab = "", ylab = "e0")
lines(e0_norway[, year], e0_norway[, e0F], col = "blue")
legend("bottomright", legend = c("Japan", "Norway"), col = c("black", "blue"),
    lty = 1, bty = "n")

}
