% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpp.R
\name{get_wpp_mig}
\alias{get_wpp_mig}
\title{Country-specific WPP net migration counts}
\usage{
get_wpp_mig(country)
}
\arguments{
\item{country}{Name of country.}
}
\value{
\code{data.table} object with columns \code{year} and \code{mig}.
}
\description{
The function returns net migration totals over time for a specific country
      from the UN's World Population Prospects.
}
\details{
The data corresponds to the dataset \code{\link[wpp2022]{migration1dt}}.
}
\examples{
# Compare migration of three countries
country.colors <- list(Germany = "black", China = "orange", 
                        `United States of America` = "red")
plot(0,1, type = "n", ylim = c(-1000, 2000), xlim = c(1950, 2100), 
    main = "Net Migration", ylab = "Migration/1000", xlab = "")
for(cntry in names(country.colors)){
    mig <- get_wpp_mig(cntry)
    lines(mig[, year], mig[, mig], col = country.colors[[cntry]])
}
abline(h = 0, col = "grey")
legend("topright", legend = names(country.colors), col = unlist(country.colors),
    lty = 1, bty = "n")

}
