% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpp.R
\name{get_wpp_regions}
\alias{get_wpp_regions}
\title{WPP regions}
\usage{
get_wpp_regions(sort = TRUE)
}
\arguments{
\item{sort}{Logical. If \code{TRUE} results are sorted alphabetically.}
}
\value{
Character vector of names of regions.
}
\description{
Returns a vector of names of aggregated regions, such as continents or 
    other aggregations of countries, included in the WPP data.
}
\details{
The function uses the \code{\link[wpp2022]{UNlocations}} dataset of the
     \pkg{wpp2022} package. It selects locations where \code{location_type} is other than four,
     as four is the code for countries.
}
\examples{
regions <- get_wpp_regions()
head(regions)
tail(regions)

}
