% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpp.R
\name{get_wpp_tfr}
\alias{get_wpp_tfr}
\title{Country-specific WPP total fertility rate}
\usage{
get_wpp_tfr(country)
}
\arguments{
\item{country}{Name of country.}
}
\value{
\code{data.table} object with columns \code{year} and \code{tfr}.
}
\description{
The function returns the total fertility rate over time for a specific country
      from the UN's World Population Prospects.
}
\details{
The data is extracted by combining datasets \code{\link[wpp2022]{tfr1dt}} and
     \code{\link[wpp2022]{tfrproj1dt}}.
}
\examples{
# Compare TFR of Niger and Brazil
tfr_niger <- get_wpp_tfr("Niger")
tfr_brazil <- get_wpp_tfr("Brazil")
plot(tfr_niger[, year], tfr_niger[, tfr], type = "l", ylim = c(0,8),
    main = "TFR of Niger and Brazil", xlab = "", ylab = "TFR")
lines(tfr_brazil[, year], tfr_brazil[, tfr], col = "blue")
legend("bottomleft", legend = c("Niger", "Brazil"), col = c("black", "blue"),
    lty = 1, bty = "n")

}
