% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{load_forecast}
\alias{load_forecast}
\title{Load Existing Population Forecast}
\usage{
load_forecast(country, simulation_name, output_dir = tempdir(), units = 1000)
}
\arguments{
\item{country}{Name of country.}

\item{simulation_name}{Character string to identify the forecast to load. It is used to 
construct the name of the simulation directory from which the prediction will be extracted.}

\item{output_dir}{Base directory where the prediction object is assumed to be stored. The full name of the 
simulation directory is  constructed as \code{output_dir/simulation_name}.}

\item{units}{Scale of the output data. The same meaning as in \code{\link{run_forecast}}.}
}
\description{
Loads forecast results from a directory given by name
    and returns the same datasets as the \code{\link{run_forecast}} function.
}
\examples{
# Run a forecast (a random simulation name will be assigned)
forecast <- run_forecast("Argentina", start_year = 2024, keep_sim_dir = TRUE)

# Load forecast of that name
loaded_forecast <- load_forecast("Argentina", simulation_name = forecast$simulation_name,
                                 output_dir = forecast$output_dir)

# They should be equal                                 
identical(forecast, loaded_forecast)

# Remove the forecast
remove_forecast(forecast)
remove_forecast(loaded_forecast) # nothing to remove
}
