% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{move_pop}
\alias{move_pop}
\title{Moving Population Counts to Specified Date}
\usage{
move_pop(pop, country, initial_date, desired_date = NULL, year = 2023)
}
\arguments{
\item{pop}{Data table with population data in single year ages. 
It should have a column \dQuote{age}, \dQuote{popM} (male population) and 
\dQuote{popF} (female population).}

\item{country}{Name of the country to which the data belong to. It is used to 
extract other demographic indicators (fertility, mortality, migration).}

\item{initial_date}{Character string of the form \dQuote{YYYY-MM-DD}, specifying the date
assigned to the given population data \code{pop}.}

\item{desired_date}{Character string of the form \dQuote{YYYY-MM-DD}, specifying the date
to which the \code{pop} dataset should be moved to. By default,
it is December 31 of the year given by the argument \code{year}.}

\item{year}{Integer specifying the year of the population counts. Together 
with the \code{country} argument it is used for extracting vital rates from WPP.}
}
\value{
Data table with columns \dQuote{age}, \dQuote{popM} and \dQuote{popF} 
    in which the population counts have been shifted to the desired date.
}
\description{
Shifts given population from a given date to a later date.
}
\details{
The function extracts age-specific fertility, mortality and net migration
    for the given country and year from the WPP datasets and shifts the given population counts 
    to the desired date by performing population projections using the exponential growth method.
}
\examples{

# extract 1-year population of Sweden in 2024
pop_start <- get_wpp_pop("Sweden", n = 1, year = 2024)

# pretend this data is from January 1 2025
# -> shift to December 31 (default desired_date)
pop_end <- move_pop(pop_start, "Sweden", "2025-01-01", year = 2025)

# compare with WPP projection from Dec 31, 2025
pop_wpp <- get_wpp_pop("Sweden", n = 1, year = 2025)

res <- data.frame(start = pop_start[, sum(popM + popF)],
                  end = pop_end[, sum(popM + popF)],
                  wpp = pop_wpp[, sum(popM + popF)])
print(res)

}
