% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{reduce_oag}
\alias{reduce_oag}
\title{Collapsing Old Ages}
\usage{
reduce_oag(pop, oag_new, pop_columns = NULL)
}
\arguments{
\item{pop}{Data table with columns \dQuote{age} and other columns containing 
age-specific population data. It can be either for single year or five year age groups.}

\item{oag_new}{Integer defining the new OAG.It should be smaller
 than the current OAG, i.e. the maximum age available in the \code{pop} dataset.
 If \code{pop} is defined for five year age groups,
\code{oag_new} must be a multiple of 5.}

\item{pop_columns}{Character vector of columns of the \code{pop} dataset to collapse. 
By default, all columns that are not called \dQuote{age} are used.}
}
\value{
Data table with the same columns as in the \code{pop} dataset 
    (or defined by \code{pop_columns}), where all age groups equal to and older than 
    \code{oag_new} are aggregated.
}
\description{
Collapse old ages into a new open age group (OAG).
}
\examples{
# extract population of Canada in 1970
pop_full <- get_wpp_pop("Canada", year = 1970)

# sum population by sex
pop_full[, pop_tot := popM + popF]

# collapse into OAG of 95
pop_reduced <- reduce_oag(pop_full, oag_new = 95)

tail(pop_full)
tail(pop_reduced)

# works for 5-year age groups as well
pop5_full <- get_wpp_pop("Canada", n = 5)
pop5_reduced <- reduce_oag(pop5_full, oag_new = 80)

tail(pop5_full)
tail(pop5_reduced)

}
