% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{graduate_pop}
\alias{graduate_pop}
\title{Graduating Population from Five- to Single-Year Ages}
\usage{
graduate_pop(pop, method = "beers(ord)")
}
\arguments{
\item{pop}{Data table with population data aggregated to 5-year age groups. 
Columns are the same as in \code{\link{get_wpp_pop}}, i.e. 
\code{age}, \code{popF} and \code{popM}.}

\item{method}{Method to use in the \code{\link[DemoTools]{graduate}} function of the \pkg{DemoTools} package.}
}
\value{
Data table with the same columns as in the \code{pop} input argument, but 
    with age groups (rows) disaggregated into single-year ages.
}
\description{
The function disaggregates the input population from 5-year to 1-year ages.
}
\details{
The function applies the \pkg{DemoTools} function \code{\link[DemoTools]{graduate}}
    on both population columns of the input argument \code{pop} and returns the disaggregated dataset.
}
\examples{
# extract 5-year population of France in 2024 (default)
pop5 <- get_wpp_pop("France", n = 5)

# disaggregate into single-year of age
pop1 <- graduate_pop(pop5)

# plot the average  of the 5-year pop and showing it
# at the middle points of the age groups
ages5 <- seq(2, length = nrow(pop5), by = 5)
plot(ages5, pop5[, popF + popM]/5, type = "l", col = "blue",
    main = "Population of France in 2024",
    xlab = "age", ylab = "Population (in thousands)")
lines(pop1[, age], pop1[, popF + popM], col = "red")
legend("bottomleft", legend = c("original / 5", "graduated"),
    bty = "n", lty = 1, col = c("blue", "red"))

}
