% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{run_forecast}
\alias{run_forecast}
\title{Run Population Forecast}
\usage{
run_forecast(
  country,
  start_year = 2021,
  end_year = 2100,
  pop = NULL,
  tfr = NULL,
  units = 1000
)
}
\arguments{
\item{country}{Name of country.}

\item{start_year}{Start of the projection. It is the last observed year; the first projected year
is \code{start_year + 1}.}

\item{end_year}{The last year of the projection.}

\item{pop}{Data table with the initial population data at \code{start_year}.
Columns are the same as in \code{\link{get_wpp_pop}}. By default, the WPP data is used.}

\item{tfr}{Data table containing the TFR scenario. It has the same structure as
the dataset returned by \code{\link{get_wpp_tfr}}. By default the UN data is used.}

\item{units}{Scale of the output data. By default a simulation results are given in thousands.
If for example one would like to see results in millions, set \code{unit = 1e6}.}
}
\value{
List with elements:
\describe{
    \item{prediction}{\code{\link[bayesPop]{bayesPop.prediction}} object.
         (This is for debugging and will be removed in the future.)}
    \item{population_by_age_and_sex}{Data table with age- and sex-specific population by time
         (columns \code{year}, \code{age}, \code{popM}, \code{popF}).}
     \item{population_by_broad_age_group}{Data table where population is grouped by broader age groups.
         (columns \code{year}, \code{age}, \code{pop}, \code{pop_percent}).}
     \item{population_by_time}{Data table where with total population as well as by broader age groups.
         It also contains the UN median (columns \code{year}, \code{age}, \code{pop}, \code{un_pop_median}.)}
}
}
\description{
Runs forecast and returns various datasets with results.
}
\details{
For now the function only runs when \code{start_year} is 2021 or smaller.
    Note that for now each run of this function creates a new temporary directory.
    It can be removed by running \code{remove_forecast()} as in the example below.
}
\examples{
# Change the TFR of Niger to stay at the 4.0 level from 2054 onwards
my_tfr <- get_wpp_tfr("Niger")
my_tfr[year >= 2054, tfr := 4]

forecast <- run_forecast("Niger", start_year = 2021, tfr = my_tfr)

# extract total population
respop <- forecast$population_by_time[age == "0+"]

# the scenario population should be larger than UN starting in 1954
plot(respop$year, respop$pop, type = "l", col = "red",
    main = "Niger population", ylab = "Population", xlab = "")
lines(respop$year, respop$un_pop_median, col = "blue")
legend("topleft", legend = c("my scenario", "UN"), col = c("red", "blue"),
    lty = 1, bty = "n")

# remove prediction directory
remove_forecast(forecast)

}
