% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{run_forecast}
\alias{run_forecast}
\title{Run Population Forecast}
\usage{
run_forecast(
  country,
  start_year = 2021,
  end_year = 2100,
  pop = NULL,
  tfr = NULL,
  units = 1000
)
}
\arguments{
\item{country}{Name of country.}

\item{start_year}{Start of the projection. It is the last observed year; the first projected year
is \code{start_year + 1}.}

\item{end_year}{The last year of the projection.}

\item{pop}{Data table with the initial population data at \code{start_year}.
Columns are the same as in \code{\link{get_wpp_pop}}. By default, the WPP data is used.}

\item{tfr}{Data table containing the TFR scenario. It has the same structure as
the dataset returned by \code{\link{get_wpp_tfr}}. By default the UN data is used.}

\item{units}{Scale of the output data. By default a simulation results are given in thousands.
If for example one would like to see results in millions, set \code{unit = 1e6}.}
}
\value{
List with elements:
\describe{
    \item{prediction}{\code{\link[bayesPop]{bayesPop.prediction}} object.
         (This is for debugging and will be removed in the future.)}
    \item{population_by_age_and_sex}{Data table with age- and sex-specific population by time
         (columns \code{year}, \code{age}, \code{popM}, \code{popF}).}
     \item{population_by_broad_age_group}{Data table where population is grouped by broader age groups.
         (columns \code{year}, \code{age}, \code{pop}, \code{pop_percent}).}
     \item{population_by_time}{Data table where with total population as well as by broader age groups.
         It also contains the UN median and the 95\% probability intervals 
         (columns \code{year}, \code{age}, \code{pop}, \code{un_pop_median}, 
         \code{un_pop_95low}, \code{un_pop_95high}).}
     \item{tfr_by_time}{Data table with the total fertility forecast and the UN median and the 
         95\% probability intervals (columns \code{year}, \code{tfr}, \code{un_tfr_median}, 
         \code{un_tfr_95low}, \code{un_tfr_95high}).}
     \item{annual_growth_rate}{Data table with the forecast of annual growth rate as a percentage,
             computed as \eqn{log(P_t/P_{t-1})*100}, for the same age groups as in \code{population_by_time}
             (columns \code{year}, \code{age}, \code{growth_rate}).}
     \item{births_counts_rates}{Data table with the forecast of total births (column \code{births}),
             crude birth rate (\code{cbr}) and the UN median and 95\% probability intervals 
             of the respective quantities
             (columns \code{un_births_median}, \code{un_births_95low}, \code{un_births_95high}, 
             \code{un_cbr_median}, \code{un_cbr_95low}, \code{un_cbr_95high}).}
     \item{deaths_counts_rates}{Data table with the forecast of total deaths (column \code{deaths}),
             crude death rate (\code{cdr}) and the UN median and 95\% probability intervals 
             of the respective quantities
             (columns \code{un_deaths_median}, \code{un_deaths_95low}, \code{un_deaths_95high}, 
             \code{un_cdr_median}, \code{un_cdr_95low}, \code{un_cdr_95high}).}
     \item{yadr}{Data table with the forecast of the young age dependency ratio [(0-19)/(20-64)] (column \code{yadr})
             and the UN median and 95\% probability intervals of the same quantity (columns 
             \code{un_yadr_median}, \code{un_yadr_95low}, \code{un_yadr_95high}).}
     \item{oadr}{Data table with the forecast of the old age dependency ratio [65+/(20-64)] (column \code{oadr})
             and the UN median and 95\% probability intervals of the same quantity (columns 
             \code{un_oadr_median}, \code{un_oadr_95low}, \code{un_oadr_95high}).}
     \item{pop_aging_and_pop_size}{Data table with the forecast of percent population of 65+ (column \code{percent65}),
             total population (column \code{pop}) and the UN median of the same 
             two quantities (columns \code{un_pop_median}, \code{un_percen65}).}
     \item{cdr_by_e0}{Data table with the forecast of the crude death rate (\code{cdr}), life expecancy at birth
             (\code{e0}) and the UN median of the same two quantities (columns \code{un_cdr_median}, \code{un_e0_median}).}
     \item{cbr_by_tfr}{Data table with the forecast of the crude birth rate (\code{cbr}), total fertility rate
             (\code{tfr}) and the UN median of the same two quantities (columns \code{un_cbr_median}, \code{un_tfr_median}).}
     
}
}
\description{
Runs forecast and returns various datasets with results.
}
\details{
For now the function only runs when \code{start_year} is 2021 or smaller.
    Note that for now each run of this function creates a new temporary directory.
    It can be removed by running \code{remove_forecast()} as in the example below.
}
\examples{
# Change the TFR of Niger to stay at the 4.0 level from 2054 onwards
my_tfr <- get_wpp_tfr("Niger")
my_tfr[year >= 2054, tfr := 4]

forecast <- run_forecast("Niger", start_year = 2021, tfr = my_tfr)

# plot population and fertility
################################
par(mfrow = c(1,2))

# extract total population
respop <- forecast$population_by_time[age == "Total"]

# the scenario population should be larger than UN starting in 1954
plot(respop$year, respop$pop, type = "l", col = "red",
    main = "Niger population", ylab = "Population", xlab = "")
lines(respop$year, respop$un_pop_median, col = "blue")
lines(respop$year, respop$un_pop_95low, col = "blue", lty = 2)
lines(respop$year, respop$un_pop_95high, col = "blue", lty = 2)
legend("topleft", legend = c("my scenario", "UN", "UN 95\% PI"), col = c("red", "blue", "blue"),
    lty = c(1, 1, 2), bty = "n")
    
# extract fertility results
restfr <- forecast$tfr_by_time
plot(restfr$year, restfr$tfr, type = "l", col = "red", ylim = c(0, 8),
    main = "Niger TFR", ylab = "Total fertility rate", xlab = "")
lines(restfr$year, restfr$un_tfr_median, col = "blue")
lines(restfr$year, restfr$un_tfr_95low, col = "blue", lty = 2)
lines(restfr$year, restfr$un_tfr_95high, col = "blue", lty = 2)
legend("bottomleft", legend = c("my scenario", "UN", "UN 95\% PI"), col = c("red", "blue", "blue"),
    lty = c(1, 1, 2), bty = "n")


# remove prediction directory
remove_forecast(forecast)

}
