% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{run_forecast}
\alias{run_forecast}
\alias{remove_forecast}
\title{Run Population Forecast}
\usage{
run_forecast(
  country,
  start_year = 2023,
  end_year = 2100,
  pop = NULL,
  tfr = NULL,
  e0 = NULL,
  mig = NULL,
  units = 1000,
  output_dir = NULL,
  ...
)

remove_forecast(forecast, verbose = TRUE)
}
\arguments{
\item{country}{Name of country.}

\item{start_year}{Start of the projection. It is the last observed year; the first projected year
is \code{start_year + 1}.}

\item{end_year}{The last year of the projection.}

\item{pop}{Data table with the initial population data at \code{start_year}.
Columns are the same as in \code{\link{get_wpp_pop}}. By default, the WPP data is used.}

\item{tfr}{Data table containing the TFR scenario. It has the same structure as
the dataset returned by \code{\link{get_wpp_tfr}}. By default the UN data is used.}

\item{e0}{Data table containing the life expectancy scenario. It has the same structure as
the dataset returned by \code{\link{get_wpp_e0}}. By default the UN data is used.}

\item{mig}{Data table containing the migration scenario. It has the same structure as
the dataset returned by \code{\link{get_wpp_mig}}. By default the UN data is used.}

\item{units}{Scale of the output data. By default a simulation results are given in thousands.
If for example one would like to see results in millions, set \code{unit = 1e6}.}

\item{output_dir}{Directory where the prediction object should be stored. If \code{NULL} 
a temporary directory is used which is removed after the simulation is finished.}

\item{\dots}{Additional arguments passed to the \code{\link[bayesPop]{pop.predict}} 
function of \pkg{bayesPop}.}

\item{forecast}{List returned by \code{run_forecast}}

\item{verbose}{Logical that switches log messages on and off.}
}
\value{
List with elements:
\describe{
    \item{population_by_age_and_sex}{Data table with age- and sex-specific population by time
         (columns \code{year}, \code{age}, \code{popM}, \code{popF}).}
     \item{population_by_broad_age_group}{Data table where population is grouped by broader age groups.
         (columns \code{year}, \code{age}, \code{pop}, \code{pop_percent}).}
     \item{population_by_time}{Data table where with total population as well as by broader age groups.
         It also contains the UN median and the 95\% probability intervals 
         (columns \code{year}, \code{age}, \code{pop}, \code{un_pop_median}, 
         \code{un_pop_95low}, \code{un_pop_95high}).}
     \item{tfr_by_time}{Data table with the total fertility forecast and the UN median and the 
         95\% probability intervals (columns \code{year}, \code{tfr}, \code{un_tfr_median}, 
         \code{un_tfr_95low}, \code{un_tfr_95high}).}
     \item{e0_by_time}{Data table with the forecast of sex specific life expectancy at birth as well as the UN medians,
          including the 95\% probability intervals (columns \code{year}, \code{sex}, \code{e0}, \code{un_e0_median}, 
         \code{un_e0_95low}, \code{un_e0_95high}).}
     \item{annual_growth_rate}{Data table with the forecast of annual growth rate as a percentage,
             computed as \eqn{log(P_t/P_{t-1})*100}, for the same age groups as in \code{population_by_time}
             (columns \code{year}, \code{age}, \code{growth_rate}).}
     \item{births_counts_rates}{Data table with the forecast of total births (column \code{births}),
             crude birth rate (\code{cbr}) and the UN median and 95\% probability intervals 
             of the respective quantities
             (columns \code{un_births_median}, \code{un_births_95low}, \code{un_births_95high}, 
             \code{un_cbr_median}, \code{un_cbr_95low}, \code{un_cbr_95high}).}
     \item{deaths_counts_rates}{Data table with the forecast of total deaths (column \code{deaths}),
             crude death rate (\code{cdr}) and the UN median and 95\% probability intervals 
             of the respective quantities
             (columns \code{un_deaths_median}, \code{un_deaths_95low}, \code{un_deaths_95high}, 
             \code{un_cdr_median}, \code{un_cdr_95low}, \code{un_cdr_95high}).}
     \item{yadr}{Data table with the forecast of the young age dependency ratio [(0-19)/(20-64) * 100] (column \code{yadr})
             and the UN median and 95\% probability intervals of the same quantity (columns 
             \code{un_yadr_median}, \code{un_yadr_95low}, \code{un_yadr_95high}).}
     \item{oadr}{Data table with the forecast of the old age dependency ratio [65+/(20-64) * 100] (column \code{oadr})
             and the UN median and 95\% probability intervals of the same quantity (columns 
             \code{un_oadr_median}, \code{un_oadr_95low}, \code{un_oadr_95high}).}
     \item{pop_aging_and_pop_size}{Data table with the forecast of percent population of 65+ (column \code{percent65}),
             total population (column \code{pop}) and the UN median of the same 
             two quantities (columns \code{un_pop_median}, \code{un_percen65}).}
     \item{cdr_by_e0}{Data table with the forecast of the crude death rate (\code{cdr}), life expecancy at birth
             (\code{e0}) and the UN median of the same two quantities (columns \code{un_cdr_median}, \code{un_e0_median}).}
     \item{cbr_by_tfr}{Data table with the forecast of the crude birth rate (\code{cbr}), total fertility rate
             (\code{tfr}) and the UN median of the same two quantities (columns \code{un_cbr_median}, \code{un_tfr_median}).}
     \item{prediction}{\code{\link[bayesPop]{bayesPop.prediction}} object. Only returned if \code{output_dir}
             is not \code{NULL}.}
     
}
}
\description{
Runs forecast and returns various datasets with results.
}
\details{
The \code{run_forecast} function launches population prediction for the given country via the \code{\link[bayesPop]{pop.predict}} 
    function from the \pkg{bayesPop} package. It uses default input datasets from the \pkg{wpp2022} package, 
    while possibly overwriting the initial population (if given in the argument \code{pop}), 
    the total fertility rate (argument \code{tfr}), the sex-specific life expectancy at birth
    (argument \code{e0}) or/and the total net migration (argument \code{mig}). 
    If \code{e0} is not supplied, the \code{\link[bayesPop]{pop.predict}} function is called
    with the argument \code{fixed.mx = FALSE}, so that projected mortality rates from the \pkg{wpp2022} package
    are used instead of the life expectancy at birth. Currently, it is assumes that it is a deterministic 
    projection, with one trajectory of all, TFR, life expectancy and migration.
    
    If it is desired to preserve the resulting 
    \code{\link[bayesPop]{bayesPop.prediction}} object, argument \code{output_dir} should be given. 
    In such a case, the prediction object is returned. Otherwise a temporary directory is used and
    it is deleted after results are collected. For a manual deletion of the simulation directory, one can use
    the function \code{remove_forecast()} while passing the object returned by \code{run_forecast}.
}
\examples{
# Change the TFR of Niger to stay at the 4.0 level from 2054 onwards
my_tfr <- get_wpp_tfr("Niger")
my_tfr[year >= 2054, tfr := 4]

# Decrease e0  by 4 (male) and 3 (female) for the same time range 
my_e0 <- get_wpp_e0("Niger")
my_e0[year >= 2054, `:=`(e0M = e0M - 4, e0F = e0F - 3)]

forecast <- run_forecast("Niger", start_year = 2030, tfr = my_tfr, e0 = my_e0)

# plot population, fertility, and life expectancy
#################################################
par(mfrow = c(2,2))

# extract total population
respop <- forecast$population_by_time[age == "Total"]

# the scenario population should be larger than UN starting in 1954
plot(respop$year, respop$pop, type = "l", col = "red",
    main = "Niger population", ylab = "Population", xlab = "")
lines(respop$year, respop$un_pop_median, col = "blue")
lines(respop$year, respop$un_pop_95low, col = "blue", lty = 2)
lines(respop$year, respop$un_pop_95high, col = "blue", lty = 2)
legend("topleft", legend = c("my scenario", "UN", "UN 95\% PI"), col = c("red", "blue", "blue"),
    lty = c(1, 1, 2), bty = "n")
    
# extract fertility
restfr <- forecast$tfr_by_time
plot(restfr$year, restfr$tfr, type = "l", col = "red", ylim = c(0, 8),
    main = "Niger TFR", ylab = "Total fertility rate", xlab = "")
lines(restfr$year, restfr$un_tfr_median, col = "blue")
lines(restfr$year, restfr$un_tfr_95low, col = "blue", lty = 2)
lines(restfr$year, restfr$un_tfr_95high, col = "blue", lty = 2)
legend("bottomleft", legend = c("my scenario", "UN", "UN 95\% PI"), col = c("red", "blue", "blue"),
    lty = c(1, 1, 2), bty = "n")
    
# extract crude birth rate
rescbr <- forecast$births_counts_rates
plot(rescbr$year, rescbr$cbr, type = "l", col = "red", ylim = c(10,60),
    main = "Niger CBR", ylab = "Crude Birth Rate", xlab = "")
lines(rescbr$year, rescbr$un_cbr_median, col = "blue")
lines(rescbr$year, rescbr$un_cbr_95low, col = "blue", lty = 2)
lines(rescbr$year, rescbr$un_cbr_95high, col = "blue", lty = 2)
legend("bottomleft", legend = c("my scenario", "UN", "UN 95\% PI"), col = c("red", "blue", "blue"),
    lty = c(1, 1, 2), bty = "n")

# extract life expectancy at birth (all sexes)
rese0 <- forecast$e0_by_time[sex == "B"]
plot(rese0$year, rese0$e0, type = "l", col = "red", 
    ylim = range(rese0$un_e0_median, rese0$un_e0_95low, rese0$un_e0_95high, na.rm = TRUE),
    main = "Niger's total e0", ylab = "Life expectancy at birth", xlab = "")
lines(rese0$year, rese0$un_e0_median, col = "blue")
lines(rese0$year, rese0$un_e0_95low, col = "blue", lty = 2)
lines(rese0$year, rese0$un_e0_95high, col = "blue", lty = 2)
legend("bottomleft", legend = c("my scenario", "UN", "UN 95\% PI"), col = c("red", "blue", "blue"),
    lty = c(1, 1, 2), bty = "n")

}
